+++ PSFramework/changelog.md
@@ -1,4 +1,17 @@
 ﻿# CHANGELOG
+
+## ???
+
+- Upd: Logging Component - Disabled wait time in logging cycle if messages pending, to avoid delays during message floods
+- Fix: Logging Component - Logging Provider Instances would ignore updating filters at runtime
+- Fix: Logging Component - Execute proper cleanup when provider / instance get disabled explicitly
+
+## 1.4.143-preview3 : 2020-08-29
+
+- New: Added Tab Expansion Plus Plus code to provide extended tab completion on PS3-4
+- New: Argument Transformation Attribute: PsfDynamicTransform - allows dynamic object conversion from PSObject, hashtable, or type from a different library version
+- Upd: Invoke-PSFProtectedCommand now respects explicitly bound `-WhatIf` and `-Confirm` parameters.
+
 ## 1.3.140-preview2 : 2020-08-20
 
 - Upd: Logging Provider logfile - added configuration for encoding